requirements.txt
numpy==1.19.2
Faker==4.14.0
vega_datasets==0.9.0
seaborn==0.11.1
wordcloud==1.8.1
altair==4.1.0
streamlit==0.74.1
requests==2.25.1
PuLP==2.3.1
psycopg2==2.8.6
scikit_learn==0.24.1
pytz==2020.1
timezonefinder==5.2.0
geopandas==0.6.1
descartes==1.1.0
matplotlib==3.3.4
